some updates

+++ Backend/RRSS/src/main/java/com/demo/rrss/rrssbackend/service/ForumService.java
@@ -1,11 +1,7 @@
 package com.demo.rrss.rrssbackend.service;
 
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,6 +11,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import com.demo.rrss.rrssbackend.entity.Forum;
 import com.demo.rrss.rrssbackend.repository.ForumRepository;
+import com.demo.rrss.rrssbackend.repository.UsersRepository;
 import com.demo.rrss.rrssbackend.rest.request.ForumRequest;
 
 
@@ -22,13 +19,18 @@ import com.demo.rrss.rrssbackend.rest.request.ForumRequest;
 public class ForumService {
     @Autowired
 	ForumRepository repository;
-
+	@Autowired
+	UsersRepository userRepository;
 	public Forum getForum(Long forumId) {
 		Optional<Forum> response = repository.findById(forumId);
 		return response.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Forum not found"));
 	}
 
-	public void addForum(ForumRequest request) {
+	public void addForum(ForumRequest request, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if(!userRepository.findById(userId).get().getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You do not have permission to do that!");
+		}
 		Forum forum = new Forum();
 		forum.setForumName(request.getForumName());
 		forum.setForumDescription(request.getForumDescription());
@@ -36,7 +38,11 @@ public class ForumService {
 		repository.save(forum);
 	}
 
-	public void updateForum(Long forumId, ForumRequest request) {
+	public void updateForum(Long forumId, ForumRequest request, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if(!userRepository.findById(userId).get().getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You do not have permission to do that!");
+		}
 		Optional<Forum> existingForum = repository.findById(forumId);
 		if (existingForum.isPresent()){
 			Forum forum = existingForum.get();
@@ -49,14 +55,23 @@ public class ForumService {
 		}
 	}
 
-	public void deleteForum(Long forumId) {
+	public void deleteForum(Long forumId, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if(!userRepository.findById(userId).get().getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You do not have permission to do that!");
+		}
 		if (repository.existsById(forumId))
 			repository.deleteById(forumId);
 		else
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Forum not found");
     }
 
-	public List<Forum> getAllForums() {
+	public List<Forum> getAllForums(Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if(!userRepository.findById(userId).get().getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You do not have permission to do that!");
+		}
+
 		return (List<Forum>) repository.findAll();
 	}
